initialize row set absraction
