+++ tests/integration/dbapi/async/V2/test_system_engine_async.py
@@ -4,7 +4,8 @@ from typing import List
 from pytest import raises
 
 from firebolt.async_db import Connection
-from firebolt.common._types import ColType, Column
+from firebolt.common._types import ColType
+from firebolt.common.row_set.types import Column
 from firebolt.utils.exception import FireboltStructuredError
 from tests.integration.dbapi.utils import assert_deep_eq
