+++ tests/integration/dbapi/conftest.py
@@ -5,7 +5,8 @@ from typing import List
 
 from pytest import fixture
 
-from firebolt.common._types import STRUCT, ColType, Column
+from firebolt.common._types import STRUCT, ColType
+from firebolt.common.row_set.types import Column
 from firebolt.db import ARRAY, DECIMAL, Connection
 
 LOGGER = getLogger(__name__)
