+++ tests/unit/db/test_cursor.py
@@ -6,9 +6,10 @@ from httpx import URL, HTTPStatusError, Request, StreamError, codes
 from pytest import LogCaptureFixture, mark, raises
 from pytest_httpx import HTTPXMock
 
-from firebolt.common._types import Column
+from firebolt.common.constants import CursorState
+from firebolt.common.row_set.types import Column
 from firebolt.db import Cursor
-from firebolt.db.cursor import ColType, CursorState
+from firebolt.db.cursor import ColType
 from firebolt.utils.exception import (
     ConfigurationError,
     CursorClosedError,
